feat(260607-l9m): widen ReceiptSummary numeric fields int->float64
Production failed on 2026-06-07 with:
  cannot unmarshal number 85.56 into Go struct field
  ReceiptSummary.summary.total_units of type int

Extends the 260607-k1n int->float64 widening pattern from
ReceiptItem.Quantity to ReceiptSummary.TotalUnits + .TotalCases.

- types.go: TotalUnits/TotalCases int -> float64 with doc comment
  pointing at the validate.go rounding boundary
- validate.go Check 3: round summary side via int(math.Round(...)) at
  the comparison boundary, mirroring the existing totalQty rounding
- parser_test.go: %d -> %v format verb fix for TotalUnits assertion
  (would emit %!d(float64=40) at runtime after widening)

Caller audit:
- inventory.PendingPurchase.TotalUnits/.TotalCases stay *int — DB column
  is INTEGER and ReceiptSummary is never persisted directly. Out of
  scope per the user constraint (deferred follow-up in SUMMARY).
- worker.go has no .TotalUnits/.TotalCases references; only persists
  summary.Vendor/.Tax/.Total.
- worker_test.go's 5 ReceiptSummary literals use untyped int (1, 0)
  which assign cleanly to float64 per Go spec — no change needed.

go vet ./internal/receipt/... clean; go build ./... clean;
TestParseJSONBody_* + TestValidateReceiptData* +
TestRunIngestCycle_ScenarioTable all pass.

+++ backend/internal/receipt/types.go
@@ -49,10 +49,18 @@ type ReceiptItem struct {
 }
 
 // ReceiptSummary is the summary block parsed from a receipt.
+//
+// TotalUnits/TotalCases are float64 so JSON values like total_units: 85.56
+// (which Anthropic returns for some receipts — e.g. the 2026-06-07
+// Restaurant Depot case, 260607-l9m) unmarshal cleanly instead of failing
+// the strict int decoder. The DB columns pending_purchases.total_units /
+// total_cases are INTEGER, but ReceiptSummary is NEVER persisted directly;
+// worker.go only writes summary.Vendor/.Tax/.Total. The validate.go
+// comparison rounds via int(math.Round(...)) at the boundary.
 type ReceiptSummary struct {
 	Vendor     string  `json:"vendor"`
-	TotalUnits int     `json:"total_units"`
-	TotalCases int     `json:"total_cases"`
+	TotalUnits float64 `json:"total_units"`
+	TotalCases float64 `json:"total_cases"`
 	Tax        float64 `json:"tax"`
 	Total      float64 `json:"total"`
 }
